backend/requirements.txt
django>=5.0,<6.0
bcrypt>=4.1,<5.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
psycopg2-binary>=2.9,<3.0
celery[redis]>=5.3,<6.0
redis>=5.0,<6.0
python-dotenv>=1.0,<2.0
gunicorn>=21.2,<23.0
openpyxl>=3.1,<4.0
weasyprint>=62.0,<64.0
ruff>=0.4,<1.0
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
Pillow>=10.0,<12.0
boto3>=1.34,<2.0
